Docker tags update - 1
Wrote a Python script to replace the tags in the Dockerfiles. The script accepts two command-line arguments - file path where replacement to be made and the latest image tag.

Added a job in the YAML file to run this python script to replace the tags in the Dockerfiles.
Additionally, it also contains steps to automatically add, commit, push changes to the files in the repo.